DPG-034: Introduce ProfilesRepository and refactor CLI to use it
