+++ test/profiles-file.test.js
@@ -1,10 +1,6 @@
 import { describe, it, expect } from 'vitest'
-import { tmpdir } from 'node:os'
-import fs from 'node:fs/promises'
-import { mkdtemp } from 'node:fs/promises'
 import path from 'node:path'
-import {resolveProfilesPath, findProfileByLabel, loadProfileByLabel, loadAllProfiles} from '../src/profiles-file.js'
-import { makeProfile } from './fixtures/profiles.js'
+import {resolveProfilesPath, loadAllProfiles} from '../src/profiles-file.js'
 
 describe('resolveProfilesPath', () => {
   it('uses ~/.dpg-v2/profiles.json on macOS', () => {
@@ -44,52 +40,7 @@ describe('resolveProfilesPath', () => {
   })
 })
 
-describe('findProfileByLabel', () => {
-  it('returns the matching profile', () => {
-    const profiles = [
-      makeProfile({ label: 'github-main' }),
-      makeProfile({ label: 'gitlab-main', service: 'gitlab.com' })
-    ]
-
-    const result = findProfileByLabel(profiles, 'github-main')
-
-    expect(result.label).toBe('github-main')
-    expect(result.service).toBe('github.com')
-  })
-
-  it('throws if profile is not found', () => {
-    const profiles = [makeProfile({ label: 'github-main' })]
-
-    expect(() => findProfileByLabel(profiles, 'missing'))
-      .toThrow(/profile not found/i)
-  })
-
-  it('throws if profiles root is not an array', () => {
-    expect(() => findProfileByLabel({}, 'github-main'))
-      .toThrow(/array/i)
-  })
-})
-
-describe('loadProfileByLabel', () => {
-  it('loads a matching profile from disk', async () => {
-    const dir = await mkdtemp(path.join(tmpdir(), 'dpg-'))
-    const profilesPath = path.join(dir, 'profiles.json')
-
-    await fs.writeFile(
-      profilesPath,
-      JSON.stringify([
-        makeProfile({ label: 'github-main' }),
-        makeProfile({ label: 'gitlab-main', service: 'gitlab.com' })
-      ]),
-      'utf8'
-    )
-
-    const result = await loadProfileByLabel('gitlab-main', { profilesPath })
-
-    expect(result.label).toBe('gitlab-main')
-    expect(result.service).toBe('gitlab.com')
-  })
-
+describe('missing profiles file', () => {
   it('returns empty array when profiles file is missing', async () => {
     const missingPath = path.join('/definitely', 'not-there', 'profiles.json')
     const result = await loadAllProfiles({ profilesPath: missingPath })
